--- ninedot-api/src/main/java/org/sopt36/ninedotserver/auth/adapter/in/web/v1/dto/response/LoginOrSignupResponse.java
@@ -1,5 +1,0 @@
-package org.sopt36.ninedotserver.auth.adapter.in.web.v1.dto.response;
-
-public record LoginOrSignupResponse() {
-
-}
