+++ ninedot-api/src/main/java/org/sopt36/ninedotserver/auth/adapter/in/web/v1/mapper/AuthRequestMapper.java
@@ -1,7 +1,9 @@
 package org.sopt36.ninedotserver.auth.adapter.in.web.v1.mapper;
 
-import org.sopt36.ninedotserver.auth.dto.request.SignupCommand;
+import org.sopt36.ninedotserver.auth.adapter.in.web.v1.dto.request.GoogleAuthCodeRequest;
 import org.sopt36.ninedotserver.auth.adapter.in.web.v1.dto.request.SignupRequest;
+import org.sopt36.ninedotserver.auth.dto.command.GoogleLoginCommand;
+import org.sopt36.ninedotserver.auth.dto.command.SignupCommand;
 
 public class AuthRequestMapper {
 
@@ -15,6 +17,16 @@ public class AuthRequestMapper {
             apiRequest.job(),
             apiRequest.profileImageUrl(),
             apiRequest.answers()
+        );
+    }
+
+    public static GoogleLoginCommand toGoogleLoginCommand(
+        GoogleAuthCodeRequest apiRequest,
+        String clientRedirectUri
+    ) {
+        return new GoogleLoginCommand(
+            apiRequest.code(),
+            clientRedirectUri
         );
     }
